use size_t rather than unsigned int for representing buffer lengths for yajl 2.x

+++ src/api/yajl_common.h
@@ -33,6 +33,8 @@
 #ifndef __YAJL_COMMON_H__
 #define __YAJL_COMMON_H__
 
+#include <stddef.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif    
@@ -58,14 +60,14 @@ extern "C" {
 
 /** pointer to a malloc function, supporting client overriding memory
  *  allocation routines */
-typedef void * (*yajl_malloc_func)(void *ctx, unsigned int sz);
+typedef void * (*yajl_malloc_func)(void *ctx, size_t sz);
 
 /** pointer to a free function, supporting client overriding memory
  *  allocation routines */
 typedef void (*yajl_free_func)(void *ctx, void * ptr);
 
 /** pointer to a realloc function which can resize an allocation. */
-typedef void * (*yajl_realloc_func)(void *ctx, void * ptr, unsigned int sz);
+typedef void * (*yajl_realloc_func)(void *ctx, void * ptr, size_t sz);
 
 /** A structure which can be passed to yajl_*_alloc routines to allow the
  *  client to specify memory allocation functions to be used. */
