+++ src/yajl_buf.h
@@ -56,7 +56,7 @@ yajl_buf yajl_buf_alloc(yajl_alloc_funcs * alloc);
 void yajl_buf_free(yajl_buf buf);
 
 /* append a number of bytes to the buffer */
-void yajl_buf_append(yajl_buf buf, const void * data, unsigned int len);
+void yajl_buf_append(yajl_buf buf, const void * data, size_t len);
 
 /* empty the buffer */
 void yajl_buf_clear(yajl_buf buf);
@@ -65,9 +65,9 @@ void yajl_buf_clear(yajl_buf buf);
 const unsigned char * yajl_buf_data(yajl_buf buf);
 
 /* get the length of the buffer */
-unsigned int yajl_buf_len(yajl_buf buf);
+size_t yajl_buf_len(yajl_buf buf);
 
 /* truncate the buffer */
-void yajl_buf_truncate(yajl_buf buf, unsigned int len);
+void yajl_buf_truncate(yajl_buf buf, size_t len);
 
 #endif
