+++ src/yajl_lex.h
@@ -92,12 +92,12 @@ n * error messages.
  * size to get adequate performance.
  */
 yajl_tok yajl_lex_lex(yajl_lexer lexer, const unsigned char * jsonText,
-                      unsigned int jsonTextLen, unsigned int * offset,
-                      const unsigned char ** outBuf, unsigned int * outLen);
+                      size_t jsonTextLen, size_t * offset,
+                      const unsigned char ** outBuf, size_t * outLen);
 
 /** have a peek at the next token, but don't move the lexer forward */
 yajl_tok yajl_lex_peek(yajl_lexer lexer, const unsigned char * jsonText,
-                       unsigned int jsonTextLen, unsigned int offset);
+                       size_t jsonTextLen, size_t offset);
 
 
 typedef enum {
@@ -121,13 +121,13 @@ const char * yajl_lex_error_to_string(yajl_lex_error error);
 yajl_lex_error yajl_lex_get_error(yajl_lexer lexer);
 
 /** get the current offset into the most recently lexed json string. */
-unsigned int yajl_lex_current_offset(yajl_lexer lexer);
+size_t yajl_lex_current_offset(yajl_lexer lexer);
 
 /** get the number of lines lexed by this lexer instance */
-unsigned int yajl_lex_current_line(yajl_lexer lexer);
+size_t yajl_lex_current_line(yajl_lexer lexer);
 
 /** get the number of chars lexed by this lexer instance since the last
  *  \n or \r */
-unsigned int yajl_lex_current_char(yajl_lexer lexer);
+size_t yajl_lex_current_char(yajl_lexer lexer);
 
 #endif
